Appointment Booking Functionality

+++ PatientAdmin/src/controllers/SecretaryController.java
@@ -5,6 +5,7 @@
  */
 package controllers;
 import accounts.Account;
+import accounts.Admin;
 import view.SecretaryHome;
 import accounts.Secretary;
 import javax.swing.JOptionPane;
@@ -30,19 +31,41 @@ public class SecretaryController {
         view.getBtnLogout().addActionListener(e -> logout());
         view.getBtnAction().addActionListener(e -> actionRequest());
         view.getBtnDelete().addActionListener(e -> deleteRequest());
+        view.getBtnRemove().addActionListener(e -> removeAccount());
         view.getLstRequests().setListData(Secretary.getRequests());
+        view.getLstPatients().setListData(Admin.getAccounts("PATIENT"));
     }
     
     private void actionRequest() {
-        Secretary.actionRequest(view.getLstRequests().getSelectedValue());
-        JOptionPane.showMessageDialog(null, "Request Actioned");
+        
+        int alert = JOptionPane.showConfirmDialog(null, "Are you sure you want to action this request?", "Action Request", JOptionPane.YES_NO_OPTION);
+        
+        if (alert == JOptionPane.YES_OPTION) {
+            Secretary.actionRequest(view.getLstRequests().getSelectedValue());
+            JOptionPane.showMessageDialog(null, "Request Actioned");           
+        }
         view.getLstRequests().setListData(Secretary.getRequests());
     }
         
     private void deleteRequest() {
-        Secretary.deleteRequest(view.getLstRequests().getSelectedValue());
-        JOptionPane.showMessageDialog(null, "Request Deleted");
+        
+        int alert = JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this request?", "Delete Request", JOptionPane.YES_NO_OPTION);
+        
+        if (alert == JOptionPane.YES_OPTION) {
+            Secretary.deleteRequest(view.getLstRequests().getSelectedValue());
+            JOptionPane.showMessageDialog(null, "Request Deleted");            
+        }
         view.getLstRequests().setListData(Secretary.getRequests());
+    }
+    
+    private void removeAccount() {
+        int alert = JOptionPane.showConfirmDialog(null, "Are you sure you want to remove this account?", "Remove Account", JOptionPane.YES_NO_OPTION);
+        
+        if (alert == JOptionPane.YES_OPTION) {
+            Admin.removeAccount(view.getLstPatients().getSelectedValue());
+            JOptionPane.showMessageDialog(null, "Account Removed");
+        }
+        view.getLstPatients().setListData(Admin.getAccounts("PATIENT"));
     }
     
     private void logout(){
